Documents/amagara_masya/backend/amagara_masya/urls.py
from django.contrib import admin
from django.urls import path, include
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('api/core/', include('core.urls')),
    path('api/children/', include('children.urls')),
    path('api/staff/', include('staff.urls')),
    path('api/donors/', include('donors.urls')),
    path('api/reports/', include('reports.urls')),
] 