Documents/amagara_masya/backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-storages==1.14.2
boto3==1.28.62
django-allauth==0.57.0
django-crispy-forms==2.0
crispy-bootstrap5==2023.10
django-environ==0.11.2
django-phonenumber-field==7.1.0
phonenumbers==8.13.22
django-import-export==3.3.0
django-report-builder==7.0.0 